Updates To get into Working Condition
Writing to the Header File Still not Functional. Housekeeping Data needs to Be cleaned and Formatted still.

Major issue is that OBC freezes after a cetain amount of time. Not Sure What Causes this. But it seems like an eps Issue no lights are present on the Board when this happens and it ~seems~ to be variable time before it locks up